// src/app/home/navigation/types/navigation-sections.ts
export enum NavigationSections {
    Main = 'main',
    Payments = 'payments',
    PaymentsAnalytics = 'payments_analytics',
    PaymentsOperations = 'payments_operations',
    PaymentsPayouts = 'payments_payouts',
    PaymentsReports = 'payments_reports',
    PaymentsIntegrations = 'payments_integrations',
    Wallets = 'wallets',
    WalletsWallet = 'wallets_wallet',
    WalletsDeposits = 'wallets_deposits',
    WalletsWithdrawals = 'wallets_withdrawals',
    WalletsReports = 'wallets_reports',
    WalletsIntegrations = 'wallets_integrations',
    Claims = 'claims',
}